+++ examples/szconfigmanager/replace_default_config_id.py
@@ -18,7 +18,7 @@ try:
     # Create a new config.
 
     sz_config = sz_configmanager.create_config_from_config_id(current_default_config_id)
-    sz_config.add_data_source(data_source_code)
+    sz_config.register_data_source(data_source_code)
 
     # Persist the new config.